dependencies, type fixes & decorators for search api

+++ app/src/main/java/com/leandrososa/weatherapp/view/cities/CitiesIntent.kt
@@ -2,6 +2,6 @@ package com.leandrososa.weatherapp.view.cities
 
 sealed class CitiesIntent {
     data class Search(val query: String): CitiesIntent()
-    data class GoToWeather(val lat: Double, val lon: Double): CitiesIntent()
+    data class GoToWeather(val lat: Float, val lon: Float): CitiesIntent()
     object UseGeo: CitiesIntent()
 }
